5.2 Переносит рендеры в презентер
